examples/python/requirements.txt
# Core dependencies
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
zep-cloud==3.0.5

# Additional dependencies for examples
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
anthropic==0.60.0
asyncer==0.0.8
attrs==25.3.0
audioop-lts==0.2.1
backoff==2.2.1
beautifulsoup4==4.13.4
bidict==0.23.1
bs4==0.0.2
chainlit==2.6.5
charset-normalizer==3.4.2
chevron==0.14.0
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
distro==1.9.0
fastapi==0.116.1
filelock==3.18.0
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.7.0
googleapis-common-protos==1.70.0
griffe==1.9.0
grpcio==1.74.0
hf-xet==1.1.5
httpx-sse==0.4.1
huggingface-hub==0.34.3
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-text-splitters==0.3.9
langsmith==0.4.10
lazify==0.4.0
literalai==0.1.201
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.12.3
multidict==6.6.3
mypy-extensions==1.1.0
nest-asyncio==1.6.0
numpy==2.3.2
openai==1.98.0
openai-agents==0.2.4
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.1
opentelemetry-exporter-otlp-proto-common==1.34.1
opentelemetry-exporter-otlp-proto-grpc==1.34.1
opentelemetry-exporter-otlp-proto-http==1.34.1
opentelemetry-instrumentation==0.50b0
opentelemetry-proto==1.34.1
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.50b0
opentelemetry-semantic-conventions-ai==0.4.11
opentelemetry-util-http==0.50b0
orjson==3.11.1
packaging==25.0
posthog==3.25.0
propcache==0.3.2
protobuf==5.29.5
pydantic-settings==2.10.1
PyJWT==2.10.1
python-engineio==4.12.2
python-multipart==0.0.18
python-socketio==5.13.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.26.0
simple-websocket==1.1.0
soupsieve==2.7
SQLAlchemy==2.0.42
sse-starlette==3.0.2
starlette==0.47.2
syncer==2.0.3
tenacity==9.1.2
tokenizers==0.21.4
tomli==2.2.1
tqdm==4.67.1
traceloop-sdk==0.44.1
types-requests==2.32.4.20250611
typing-inspect==0.9.0
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==0.20.0
wrapt==1.17.2
yarl==1.20.1
zstandard==0.23.0
